+++ pyrosm/delta_compression.pyx
@@ -6,7 +6,8 @@ cdef delta_decode_latitude(data, node_granularity, lat_offset):
     lats_deltas = np.zeros(len(data.lat) + 1, dtype=np.int64)
     lats_deltas[1:] = list(data.lat)
     lats = (np.cumsum(lats_deltas)[1:] * node_granularity + lat_offset) / div
-    return lats
+    # Return 32 bit float (coordinate precision with 7 decimals)
+    return lats.astype(np.float32)
 
 
 cdef delta_decode_longitude(data, node_granularity, lon_offset):
@@ -14,18 +15,21 @@ cdef delta_decode_longitude(data, node_granularity, lon_offset):
     lons_deltas = np.zeros(len(data.lon) + 1, dtype=np.int64)
     lons_deltas[1:] = list(data.lon)
     lons = (np.cumsum(lons_deltas)[1:] * node_granularity + lon_offset) / div
-    return lons
+    # Return 32 bit float (coordinate precision with 7 decimals)
+    return lons.astype(np.float32)
 
 
 cdef delta_decode_id(data):
-    id_deltas = np.zeros(len(data.id) + 1, dtype=np.int64)
+    # ID needs to be 64 Bit
+    id_deltas = np.zeros(len(data.id) + 1, dtype=np.uint64)
     id_deltas[1:] = list(data.id)
     ids = np.cumsum(id_deltas)[1:]
     return ids
 
 
 cdef delta_decode_timestamp(data, timestamp_granularity):
-    timestamp_deltas = np.zeros(len(data.denseinfo.timestamp) + 1, dtype=np.int64)
+    # 32 bit unsigned integer is sufficient
+    timestamp_deltas = np.zeros(len(data.denseinfo.timestamp) + 1, dtype=np.uint32)
     timestamp_deltas[1:] = list(data.denseinfo.timestamp)
     timestamps = (np.cumsum(timestamp_deltas)[1:]
                   * timestamp_granularity / 1000).astype(int)
@@ -33,7 +37,8 @@ cdef delta_decode_timestamp(data, timestamp_granularity):
 
 
 cdef delta_decode_changeset(data):
-    changeset_deltas = np.zeros(len(data.denseinfo.changeset) + 1, dtype=np.int64)
+    # UINT 32 should be enough for Changesets for quite some time still (value between 100-200 million)
+    changeset_deltas = np.zeros(len(data.denseinfo.changeset) + 1, dtype=np.uint32)
     changeset_deltas[1:] = list(data.denseinfo.changeset)
     changesets = np.cumsum(changeset_deltas)[1:]
     return changesets
